Swap investigation guide and enrichment data places in the UI
